[MEASURE FORM API] Unit tests: added Regulations
